+++ terminal.c
@@ -1313,6 +1313,7 @@ term_init(const struct config *conf, struct fdm *fdm, struct reaper *reaper,
 #if defined(FOOT_IME_ENABLED) && FOOT_IME_ENABLED
         .ime_enabled = true,
 #endif
+        .kitty_notifications = tll_init(),
     };
 
     pixman_region32_init(&term->render.last_overlay_clip);
@@ -1817,6 +1818,13 @@ term_destroy(struct terminal *term)
         tll_remove(term->ptmx_paste_buffers, it);
     }
 
+    tll_foreach(term->kitty_notifications, it) {
+        free(it->item.id);
+        free(it->item.body);
+        free(it->item.title);
+        tll_remove(term->kitty_notifications, it);
+    }
+
     sixel_fini(term);
 
     term_ime_reset(term);
@@ -2020,6 +2028,13 @@ term_reset(struct terminal *term, bool hard)
     tll_foreach(term->alt.sixel_images, it) {
         sixel_destroy(&it->item);
         tll_remove(term->alt.sixel_images, it);
+    }
+
+    tll_foreach(term->kitty_notifications, it) {
+        free(it->item.id);
+        free(it->item.title);
+        free(it->item.body);
+        tll_remove(term->kitty_notifications, it);
     }
 
     term->grapheme_shaping = term->conf->tweak.grapheme_shaping;
@@ -3566,8 +3581,10 @@ term_bell(struct terminal *term)
         }
     }
 
-    if (term->conf->bell.notify)
-        notify_notify(term, "Bell", "Bell in terminal");
+    if (term->conf->bell.notify) {
+        notify_notify(term, "Bell", "Bell in terminal",
+                      NOTIFY_ALWAYS, NOTIFY_URGENCY_NORMAL);
+    }
 
     if (term->conf->bell.flash)
         term_flash(term, 100);